refactor: align avatar and banner settings inputs in a single row
As you can read..? This is an UI update. Not sure if it's just me but
separating the User photo and the Banner into two different "sections"
[I guess you can call it] just looks bad.

I can understand if this gets rejected, as it is just my opinion but the
User photo and the Banner is about the "avater" of the user, and just
separating them doesn't make sense (TO ME).

If you understand what I meant.